server/infra/deploy/pyproject.toml
[tool.poetry]
name = "aces-platform-algorithm-deploy"
version = "0.0.1"
description = ""
authors = []
repository = "https://github.com/aces-inc/aces-platform"

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^5.4.1"
toml = "^0.10.2"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]
boto3 = "^1.20.27"
sagemaker = "2.116.0"
protobuf = "3.20.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
